Fix tap header not showing location bar in iOS6+
For some reason, when the height is restored to the original height,
Mobile Safari pushes away the location bar. This fix doesn't restore
the faux '100%' height but it will restore when tapping on the
scroll areas. It causes a little blinking but will probably be a
rare case. Also added a condition for this code to only run on
iPhone/iPod.