Don't show subcomments for deleted parent comment
This is a pretty crude "fix" for a bug in comments where `deleted`
got propagated down to the nested partials. So if parent is
`deleted: true`, all of its child becomes deleted :(

This is supposedly a feature in Mustache and I can't control the
inheritance for the variables. There are few ways to fix this but
I just opt for this solution of not showing any child comments at
all, since I think it's pretty useless anyway.